requirements.txt
streamlit>=1.35.0
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.6.0
pytz>=2024.1
requests>=2.32.0
lxml>=5.1.0
html5lib>=1.1
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0
